+++ tools/splitobs/splitobs.c
@@ -49,7 +49,10 @@
 #define NPEY_DEFAUT        1
 #define NDIGITS_DEFAUT     4
 #define RECTANGLE_DEFAUT   {-1,-1,-1,-1,0,0,0,0}
-#define optionsDEFAUT      {"","","","","",INOUT_DEFAUT,PILOT_DEFAUT,RECTANGLE_DEFAUT,fstparam_DEFAUT,IP1_VIDE,IP1_VIDE,1,NPEX_DEFAUT,NPEY_DEFAUT,NDIGITS_DEFAUT,0,0,-1,-1}
+#define RDB_HEADER_DEFAUT     "header"
+#define RDB_DATA_DEFAUT       "data"
+#define RDB_PRIMARYKEY_DEFAUT "id_obs"
+#define optionsDEFAUT      {"" /* fstin */, "" /* obsin*/, "" /* obsout*/, "" /* gz*/, "" /* channels */, INOUT_DEFAUT /* inout */, PILOT_DEFAUT /* pilot */, RECTANGLE_DEFAUT /* rect */, fstparam_DEFAUT /* fst */, IP1_VIDE /* niveau_min */, IP1_VIDE /* niveau_max */, 1 /* channels_voulus */, NPEX_DEFAUT /* npex */, NPEY_DEFAUT /* npey */, NDIGITS_DEFAUT /* ndigits */, 0 /* check_ua4d */, 0 /* roundrobin */, -1 /* cherrypick_x */, -1 /* cherrypick_y */, RDB_HEADER_DEFAUT /* rdb_header_table */, RDB_DATA_DEFAUT /* rdb_data_table */, RDB_PRIMARYKEY_DEFAUT /* rdb_primarykey */ }
 
 /* differentes options du programme */
 #define FSTIN_OPTION       "-fstin"
@@ -61,6 +64,9 @@
 #define BURPOUT_OPTION     "-burpout" /* Remplacee par '-obsout' */
 #define ASCII_OPTION       "-ascii"   /* Remplacee par '-obsin'  */
 #define OUT_OPTION         "-out"     /* Remplacee par '-obsout' */
+#define RDB_HEADER_OPTION      "-header"
+#define RDB_DATA_OPTION        "-data"
+#define RDB_PRIMARYKEY_OPTION  "-primarykey"
 #define GZ_OPTION          "-gz"
 #define NIVEAU_MIN_OPTION  "-niveau_min"
 #define NIVEAU_MAX_OPTION  "-niveau_max"
@@ -116,19 +122,23 @@ typedef struct {
   int check_ua4d;
   int roundrobin;
   int cherrypick_x, cherrypick_y;
+  char rdb_header_table[MAXSTR], rdb_data_table[MAXSTR], rdb_primarykey[MAXSTR];
 } options, *optionsptr;
 
+typedef struct { // to be used as the argument in a callback
+  char *table_list, *primary_key;
+} sqlite_get_tables_with_primary_key_callback_arg;
 
 /*****************************************************/
 /******* Prototype des fonctions definies ************/
 /*****************************************************/
 static int sqlite_schema_callback(void *schema_void, int count, char **data, char **columns);
 static int sqlite_check_resume_callback(void *is_resume_and_rdb4_schema_present_in_DB_void, int count, char **data, char **columns);
-static int sqlite_check_tables_with_id_obs_callback(void *table_list, int count, char **data, char **columns);
-void append_id_obs_table_list_requests(char* requete_sql, char* table_list);
+static int sqlite_check_tables_with_primary_key_callback(void *callback_args, int count, char **data, char **columns);
+void append_primary_key_table_list_requests(char* requete_sql, char* table_list, char* primary_key, char* header_table);
 
 int sqlite_add_resume_request(char* obsin, char* requete_sql, char* attached_db_name);
-int sqlite_get_tables_with_id_obs(char* obsin, char* table_list);
+int sqlite_get_tables_with_primary_key(char* obsin, char* table_list, char* primary_key, char* header_table, char* data_table);
 
 int    getGZ(int iun, char* fichier, gridtype* gridptr, int niveau, float** valeurs);
 
@@ -471,9 +481,9 @@ int f77name(splitobs)(int argc, char** argv) {
      **********************************************************/
 
     strcpy(table_list,"");
-    status = sqlite_get_tables_with_id_obs(opt.obsin, table_list);
+    status = sqlite_get_tables_with_primary_key(opt.obsin, table_list, opt.rdb_primarykey, opt.rdb_header_table, opt.rdb_data_table);
     if( status != OK ) {
-      fprintf(stderr,"Fonction main: Erreur %d de la fonction sqlite_get_tables_with_id_obs pour le fichier '%s'\n", status, opt.obsin);
+      fprintf(stderr,"Fonction main: Erreur %d de la fonction sqlite_get_tables_with_primary_key pour le fichier '%s'\n", status, opt.obsin);
 
       exit_program(NOT_OK,PROGRAM_NAME,PROBLEM,VERSION);
     }
@@ -754,51 +764,65 @@ int f77name(splitobs)(int argc, char** argv) {
       if (strlen(opt.channels)==0 && opt.niveau_min == IP1_VIDE && opt.niveau_max == IP1_VIDE)
         /* Aucun filtrage vertical n'est fait */
         sprintf(requete_sql,"attach '%s' as dbin; \n"
-                "insert into header select * from dbin.header where %s(dbin.header.lat,dbin.header.lon,%d,%d,%d,%g,%g,%g,%g,%d,%d,%d,%d)=%d;\n"
-                "insert into data   select * from dbin.data   where dbin.data.id_obs in (select id_obs from header);",
-                opt.obsin, SQLFUNCTION_NAME, grid.gridid, grid.ni, grid.nj,
+                "insert into %s select * from dbin.%s where %s(dbin.%s.lat,dbin.%s.lon,%d,%d,%d,%g,%g,%g,%g,%d,%d,%d,%d)=%d;\n"
+                "insert into %s select * from dbin.%s where dbin.%s.%s in (select %s from %s);",
+                opt.obsin, opt.rdb_header_table, opt.rdb_header_table, SQLFUNCTION_NAME, opt.rdb_header_table, opt.rdb_header_table,
+                grid.gridid, grid.ni, grid.nj,
                 opt.rect.min_i, opt.rect.max_i, opt.rect.min_j, opt.rect.max_j,
                 opt.rect.min_i_equal, opt.rect.max_i_equal, opt.rect.min_j_equal, opt.rect.max_j_equal,
-                opt.inout);
+                opt.inout,
+                opt.rdb_data_table, opt.rdb_data_table, opt.rdb_data_table, opt.rdb_primarykey, opt.rdb_primarykey, opt.rdb_header_table);
       else if (strlen(opt.channels)==0 && strlen(opt.gz)==0)
         /* Le filtrage vertical est fait a l'aide d'une hauteur en pression */
         sprintf(requete_sql,"attach '%s' as dbin; \n"
-                "insert into header select * from dbin.header where %s(dbin.header.lat,dbin.header.lon,%d,%d,%d,%g,%g,%g,%g,%d,%d,%d,%d)=%d;\n"
-                "insert into data   select * from dbin.data   where dbin.data.id_obs in (select id_obs from header) and \n"
-                "  %s(dbin.data.id_obs,dbin.data.vcoord,%d,%d)=1;",
-                opt.obsin, SQLFUNCTION_NAME, grid.gridid, grid.ni, grid.nj,
+                "insert into %s select * from dbin.%s where %s(dbin.%s.lat,dbin.%s.lon,%d,%d,%d,%g,%g,%g,%g,%d,%d,%d,%d)=%d;\n"
+                "insert into %s select * from dbin.%s where dbin.%s.%s in (select %s from %s) and \n"
+                "  %s(dbin.%s.%s,dbin.%s.vcoord,%d,%d)=1;",
+                opt.obsin, opt.rdb_header_table, opt.rdb_header_table, SQLFUNCTION_NAME, opt.rdb_header_table, opt.rdb_header_table,
+                grid.gridid, grid.ni, grid.nj,
                 opt.rect.min_i, opt.rect.max_i, opt.rect.min_j, opt.rect.max_j,
                 opt.rect.min_i_equal, opt.rect.max_i_equal, opt.rect.min_j_equal, opt.rect.max_j_equal,
-                opt.inout, SQL_VERTICAL_NAME, opt.niveau_min, opt.niveau_max);
+                opt.inout,
+                opt.rdb_data_table, opt.rdb_data_table, opt.rdb_data_table, opt.rdb_primarykey, opt.rdb_primarykey, opt.rdb_header_table,
+                SQL_VERTICAL_NAME, opt.rdb_data_table, opt.rdb_primarykey, opt.rdb_data_table, opt.niveau_min, opt.niveau_max);
       else if (strlen(opt.channels)==0)
         /* Le filtrage vertical est fait a l'aide d'une hauteur en metre */
         sprintf(requete_sql,"attach '%s' as dbin; \n"
-                "insert into header select * from dbin.header where %s(dbin.header.lat,dbin.header.lon,%d,%d,%d,%g,%g,%g,%g,%d,%d,%d,%d)=%d;\n"
-                "insert into data   select data.* from dbin.data,header where dbin.data.id_obs = header.id_obs and \n"
-                "  %s(dbin.data.id_obs,header.lat,header.lon,dbin.data.vcoord+header.elev,%d,%d,%d,%d,%d)=1;",
-                opt.obsin, SQLFUNCTION_NAME, grid.gridid, grid.ni, grid.nj,
+                "insert into %s select * from dbin.%s where %s(dbin.%s.lat,dbin.%s.lon,%d,%d,%d,%g,%g,%g,%g,%d,%d,%d,%d)=%d;\n"
+                "insert into %s select %s.* from dbin.%s,%s where dbin.%s.%s = %s.%s and \n"
+                "  %s(dbin.%s.%s,%s.lat,%s.lon,dbin.%s.vcoord+%s.elev,%d,%d,%d,%d,%d)=1;",
+                opt.obsin, opt.rdb_header_table, opt.rdb_header_table, SQLFUNCTION_NAME, opt.rdb_header_table, opt.rdb_header_table,
+                grid.gridid, grid.ni, grid.nj,
                 opt.rect.min_i, opt.rect.max_i, opt.rect.min_j, opt.rect.max_j,
                 opt.rect.min_i_equal, opt.rect.max_i_equal, opt.rect.min_j_equal, opt.rect.max_j_equal,
-                opt.inout, SQL_VERTICAL_GZ_NAME, grid_gz.gridid, grid_gz.ni,
-                grid_gz.nj, opt.niveau_min, opt.niveau_max);
+                opt.inout,
+                opt.rdb_data_table, opt.rdb_data_table, opt.rdb_data_table, opt.rdb_header_table, opt.rdb_data_table, opt.rdb_primarykey, opt.rdb_header_table, opt.rdb_primarykey,
+                SQL_VERTICAL_GZ_NAME, opt.rdb_data_table, opt.rdb_primarykey, opt.rdb_header_table, opt.rdb_header_table, opt.rdb_data_table, opt.rdb_header_table,
+                grid_gz.gridid, grid_gz.ni, grid_gz.nj, opt.niveau_min, opt.niveau_max);
       else if (opt.channels_voulus==1) /* On specifie plutot les canaux voulus  */
         sprintf(requete_sql,"attach '%s' as dbin; \n"
-                "insert into header select * from dbin.header where %s(dbin.header.lat,dbin.header.lon,%d,%d,%d,%g,%g,%g,%g,%d,%d,%d,%d)=%d;\n"
-                "insert into data   select * from dbin.data   where dbin.data.id_obs in (select id_obs from header) and \n"
-                "  dbin.data.vcoord in (%s);",
-                opt.obsin, SQLFUNCTION_NAME, grid.gridid, grid.ni, grid.nj,
+                "insert into %s select * from dbin.%s where %s(dbin.%s.lat,dbin.%s.lon,%d,%d,%d,%g,%g,%g,%g,%d,%d,%d,%d)=%d;\n"
+                "insert into %s select * from dbin.%s where dbin.%s.%s in (select %s from %s) and \n"
+                "  dbin.%s.vcoord in (%s);",
+                opt.obsin, opt.rdb_header_table, opt.rdb_header_table, SQLFUNCTION_NAME, opt.rdb_header_table, opt.rdb_header_table,
+                grid.gridid, grid.ni, grid.nj,
                 opt.rect.min_i, opt.rect.max_i, opt.rect.min_j, opt.rect.max_j,
                 opt.rect.min_i_equal, opt.rect.max_i_equal, opt.rect.min_j_equal, opt.rect.max_j_equal,
-                opt.inout, opt.channels);
+                opt.inout,
+                opt.rdb_data_table, opt.rdb_data_table, opt.rdb_data_table, opt.rdb_primarykey, opt.rdb_primarykey, opt.rdb_header_table,
+                opt.rdb_data_table, opt.channels);
       else if (opt.channels_voulus==0) /* On specifie plutot les canaux exclus  */
         sprintf(requete_sql,"attach '%s' as dbin; \n"
-                "insert into header select * from dbin.header where %s(dbin.header.lat,dbin.header.lon,%d,%d,%d,%g,%g,%g,%g,%d,%d,%d,%d)=%d;\n"
-                "insert into data   select * from dbin.data   where dbin.data.id_obs in (select id_obs from header) and \n"
-                "  dbin.data.vcoord not in (%s);",
-                opt.obsin, SQLFUNCTION_NAME, grid.gridid, grid.ni, grid.nj,
+                "insert into %s select * from dbin.%s where %s(dbin.%s.lat,dbin.%s.lon,%d,%d,%d,%g,%g,%g,%g,%d,%d,%d,%d)=%d;\n"
+                "insert into %s select * from dbin.%s where dbin.%s.%s in (select %s from %s) and \n"
+                "  dbin.%s.vcoord not in (%s);",
+                opt.obsin, opt.rdb_header_table, opt.rdb_header_table, SQLFUNCTION_NAME, opt.rdb_header_table, opt.rdb_header_table,
+                grid.gridid, grid.ni, grid.nj,
                 opt.rect.min_i, opt.rect.max_i, opt.rect.min_j, opt.rect.max_j,
                 opt.rect.min_i_equal, opt.rect.max_i_equal, opt.rect.min_j_equal, opt.rect.max_j_equal,
-                opt.inout, opt.channels);
+                opt.inout,
+                opt.rdb_data_table, opt.rdb_data_table, opt.rdb_data_table, opt.rdb_primarykey, opt.rdb_primarykey, opt.rdb_header_table,
+                opt.rdb_data_table, opt.channels);
       else {
         fprintf(stderr, "Fonction main: Incapable de creer la requete SQL\n");
 
@@ -827,7 +851,7 @@ int f77name(splitobs)(int argc, char** argv) {
       // On ajoute la requete SQL pour copier les tables 'rdb4_schema' et 'resume'.
       strcat(requete_sql,sqlreq_resume);
 
-      append_id_obs_table_list_requests(requete_sql,table_list);
+      append_primary_key_table_list_requests(requete_sql,table_list,opt.rdb_primarykey,opt.rdb_header_table);
 
       strcat(requete_sql,"\ndetach dbin;");
 
@@ -937,7 +961,7 @@ int f77name(splitobs)(int argc, char** argv) {
             } /* Fin du 'if ( status != SQLITE_OK )' */
 
             status = sqlite3_exec(sqldb, sqlschema, (void*) NULL, (void*) NULL, &ErrMsg);
-            if( status != SQLITE_OK ){
+            if( status != SQLITE_OK ) {
               fprintf(stderr, "Fonction main: Erreur %d dans la fonction sqlite3_exec: %s\n", status, ErrMsg);
               if (strcmp(ErrMsg,"PRIMARY KEY must be unique")==0) {
                 fprintf(stderr,"Cette erreur est probablement due au fait que le fichier de sortie (%s) \n"
@@ -950,20 +974,23 @@ int f77name(splitobs)(int argc, char** argv) {
 
           /* On doit fabriquer la requete sql pour faire le splitting */
           sprintf(requete_sql,"drop index if exists idx1;\n"
-"PRAGMA journal_mode = OFF;\n"
-"PRAGMA  synchronous = OFF;\n"
-"attach '%s' as dbin; \n"
-"insert into header select * from dbin.header where id_obs %% %d = %d;\n"
-"insert into data   select * from dbin.data   where id_obs %% %d = %d;%s\n",
-/* "insert into header select * from dbin.header where min (  id_obs/(${maxid}/%d),%d)  = %d;\n" */
-/* "insert into data   select * from dbin.data   where min (  id_obs/(${maxid}/%d),%d)  = %d;\n" */
-/* "CREATE TABLE rdb4_schema( schema  varchar(9) );\n" */
-/* "insert into rdb4_schema values('${TYPE}');\n" */
-/* "create table resume(date integer , time integer , run varchar(9)) ;\n" */
-/* "insert into resume values(\"$DATE\",\"$HEURE\",\"$RUN\") ;\n" */
-                  opt.obsin,nsplit,id,nsplit,id,sqlreq_resume);
-          append_id_obs_table_list_requests(requete_sql,table_list);
-          strcat(requete_sql,"create index idx1 on data(id_obs,vcoord,varno);");
+                              "PRAGMA journal_mode = OFF;\n"
+                              "PRAGMA  synchronous = OFF;\n"
+                              "attach '%s' as dbin; \n"
+                              "insert into %s select * from dbin.%s where abs(%s) %% %d = %d;\n"
+                              "insert into %s select * from dbin.%s where abs(%s) %% %d = %d;%s\n",
+                  opt.obsin,
+                  opt.rdb_header_table,opt.rdb_header_table,opt.rdb_primarykey,
+                  nsplit,id,
+                  opt.rdb_data_table,opt.rdb_data_table,opt.rdb_primarykey,
+                  nsplit,id,
+                  sqlreq_resume);
+          append_primary_key_table_list_requests(requete_sql,table_list,opt.rdb_primarykey,opt.rdb_header_table);
+          if ( strcasecmp(opt.rdb_header_table,RDB_HEADER_DEFAUT) == 0   &&
+               strcasecmp(opt.rdb_data_table,RDB_DATA_DEFAUT) == 0       &&
+               strcasecmp(opt.rdb_primarykey,RDB_PRIMARYKEY_DEFAUT) == 0 ) {
+            strcat(requete_sql,"create index idx1 on data(id_obs,vcoord,varno);");
+          }
           strcat(requete_sql,"detach dbin;");
 
           printf("\nVoici la requete SQL effectuee sur la base de donnees pour creer le fichier '%s':\n",rdbout);
@@ -971,7 +998,7 @@ int f77name(splitobs)(int argc, char** argv) {
 
           /* Execution de la requete SQL sur la base de donnees finale */
           status = sqlite3_exec(sqldb, requete_sql, (void*) NULL, (void*) NULL, &ErrMsg);
-          if( status != SQLITE_OK ){
+          if( status != SQLITE_OK ) {
             fprintf(stderr, "Fonction main: Erreur %d dans la fonction sqlite3_exec: %s\n", status, ErrMsg);
             if (strcmp(ErrMsg,"PRIMARY KEY must be unique")==0) {
               fprintf(stderr,"Cette erreur est probablement due au fait que le fichier de sortie (%s) \n"
@@ -2490,25 +2517,27 @@ static int sqlite_check_resume_callback(void *is_resume_and_rdb4_schema_present_
 
 
   /***************************************************************************
-   * fonction: sqlite_get_tables_with_id_obs
+   * fonction: sqlite_get_tables_with_primary_key
    *
-   * Trouve les tables qui contiennent une colonne 'id_obs'
+   * Trouve les tables qui contiennent une colonne 'primary_key'
    *
    ***************************************************************************/
-int sqlite_get_tables_with_id_obs(char* obsin, char* table_list) {
+int sqlite_get_tables_with_primary_key(char* obsin, char* table_list, char* primary_key, char* header_table, char* data_table) {
   int status;
   char *ErrMsg;
   sqlite3 *sqldbin;
+  char sqlrequest[MAXSTR];
+  sqlite_get_tables_with_primary_key_callback_arg callback_arg;
 
   /* Cette partie sert a trouver la requete pour copier les tables 'resume' et 'rdb4_schema' */
   /* On ouvre le fichier d'input */
   status = sqlite3_open(obsin,&sqldbin);
   if ( status != SQLITE_OK ) {
-    fprintf(stderr, "Fonction sqlite_get_tables_with_id_obs: Incapable d'ouvrir le fichier '%s' avec l'erreur '%s'\n", obsin, sqlite3_errmsg(sqldbin));
+    fprintf(stderr, "Fonction sqlite_get_tables_with_primary_key: Incapable d'ouvrir le fichier '%s' avec l'erreur '%s'\n", obsin, sqlite3_errmsg(sqldbin));
 
     status = sqlite3_close(sqldbin);
     if( status != SQLITE_OK )
-      fprintf(stderr,"Fonction sqlite_get_tables_with_id_obs: Erreur %d de la fonction sqlite3_close pour le fichier '%s'\n", status, obsin);
+      fprintf(stderr,"Fonction sqlite_get_tables_with_primary_key: Erreur %d de la fonction sqlite3_close pour le fichier '%s'\n", status, obsin);
 
     return NOT_OK;
   } /* Fin du 'if ( status != SQLITE_OK )' */
@@ -2518,10 +2547,15 @@ int sqlite_get_tables_with_id_obs(char* obsin, char* table_list) {
   /* Execution de la requete SQL sur la base de donnees */
   /* Since the table 'header' and 'data' are already processed in the main */
   /*   request, we must not process them again when adding tables which */
-  /*   contains a column 'id_obs'. */
-  status = sqlite3_exec(sqldbin, "select * from sqlite_master where lower(name) not in ('header','data');", sqlite_check_tables_with_id_obs_callback, table_list, &ErrMsg);
+  /*   contains a column 'primary_key'. */
+  sprintf(sqlrequest, "select * from sqlite_master where lower(name) not in (lower('%s'),lower('%s'));", header_table, data_table);
+
+  callback_arg.table_list = table_list;
+  callback_arg.primary_key = primary_key;
+
+  status = sqlite3_exec(sqldbin,sqlrequest, sqlite_check_tables_with_primary_key_callback, &callback_arg, &ErrMsg);
   if( status != SQLITE_OK ) {
-    fprintf(stderr, "Fonction sqlite_get_tables_with_id_obs: Erreur %d pour le fichier dans la fonction sqlite3_exec: %s\n", status, ErrMsg);
+    fprintf(stderr, "Fonction sqlite_get_tables_with_primary_key: Erreur %d pour le fichier dans la fonction sqlite3_exec: %s\n", status, ErrMsg);
     sqlite3_free(ErrMsg);
 
     return NOT_OK;
@@ -2530,7 +2564,7 @@ int sqlite_get_tables_with_id_obs(char* obsin, char* table_list) {
 
   status = sqlite3_close(sqldbin);
   if( status != SQLITE_OK ) {
-    fprintf(stderr,"Fonction sqlite_get_tables_with_id_obs: Erreur %d de la fonction sqlite3_close pour le fichier '%s'\n", status, obsin);
+    fprintf(stderr,"Fonction sqlite_get_tables_with_primary_key: Erreur %d de la fonction sqlite3_close pour le fichier '%s'\n", status, obsin);
 
     return NOT_OK;
   }
@@ -2540,15 +2574,18 @@ int sqlite_get_tables_with_id_obs(char* obsin, char* table_list) {
 
 
   /***************************************************************************
-   * fonction: sqlite_check_tables_with_id_obs_callback
+   * fonction: sqlite_check_tables_with_primary_key_callback
    *
    * Cette fonction sert de 'callback' pour la requete executee par 'sqlite3_exec'.
    *    On veut savoir si elle contient les tables 'rdb4_schema' et 'resume'
    *
    ***************************************************************************/
-static int sqlite_check_tables_with_id_obs_callback(void *table_list, int count, char **data, char **columns) {
-  int idx, isTypeTable, foundID_OBS;
+static int sqlite_check_tables_with_primary_key_callback(void *void_callback_arg, int count, char **data, char **columns) {
+  int idx, isTypeTable, found_primary_key;
+  sqlite_get_tables_with_primary_key_callback_arg *callback_arg;
   char table_name[MAXSTR];
+
+  callback_arg = (sqlite_get_tables_with_primary_key_callback_arg*) void_callback_arg;
 
   isTypeTable = 0;
   for (idx = 0; idx < count; idx++) {
@@ -2559,8 +2596,10 @@ static int sqlite_check_tables_with_id_obs_callback(void *table_list, int count,
 
   if (isTypeTable == 0) return 0;
 
+  /* printf("sqlite_check_tables_with_primary_key_callback: primary_key name: '%s'\n", callback_arg->primary_key); */
+
   strcpy(table_name,"");
-  foundID_OBS=0;
+  found_primary_key=0;
   for (idx = 0; idx < count; idx++) {
     if (strcasecmp(columns[idx],"tbl_name")==0)
       strcpy(table_name,data[idx]);
@@ -2591,24 +2630,25 @@ static int sqlite_check_tables_with_id_obs_callback(void *table_list, int count,
       char regex_errbuf[MAXSTR];
       int regex_err;
 
-      regex_err = regcomp(&regex, "id_obs", REG_ICASE);
+      regex_err = regcomp(&regex, callback_arg->primary_key, REG_ICASE);
       if (regex_err!=0) {
 	regerror(regex_err, &regex, regex_errbuf, MAXSTR);
-        fprintf(stderr,"sqlite_check_tables_with_id_obs_callback: cannot compile regular expression '%s': error '%s'", "id_obs", regex_errbuf);
+        fprintf(stderr,"sqlite_check_tables_with_primary_key_callback: cannot compile regular expression '%s': error '%s'",
+                callback_arg->primary_key, regex_errbuf);
         return 1;
       }
       regex_err = regexec(&regex,data[idx],0,(regmatch_t*) NULL,0);
       if (regex_err == 0) {
         // This means that there was a match
-        foundID_OBS=1;
+        found_primary_key=1;
       }
       else if (regex_err == REG_NOMATCH) {
-        foundID_OBS=0;
+        found_primary_key=0;
       }
       else {
 	regerror(regex_err, &regex, regex_errbuf, MAXSTR);
         regfree(&regex);
-	fprintf(stderr,"sqlite_check_tables_with_id_obs_callback: Erreur '%s' avec la fonction regexec sur la ligne '%s'\n", regex_errbuf, data[idx]);
+	fprintf(stderr,"sqlite_check_tables_with_primary_key_callback: Erreur '%s' avec la fonction regexec sur la ligne '%s'\n", regex_errbuf, data[idx]);
         return 1;
       }
 
@@ -2616,32 +2656,33 @@ static int sqlite_check_tables_with_id_obs_callback(void *table_list, int count,
     }
   }
 
-  if (foundID_OBS) {
+  if (found_primary_key) {
     if (strlen(table_name)>0) {
-      if (strlen((char*) table_list)>0) {
-        strcat((char*) table_list, " ");
+      if (strlen(callback_arg->table_list)>0) {
+        strcat(callback_arg->table_list, " ");
       }
-      strcat((char*) table_list, table_name);
+      strcat(callback_arg->table_list, table_name);
+      /* printf("sqlite_check_tables_with_primary_key_callback: found table: '%s'\n", callback_arg->table_list); */
     }
     else {
-      fprintf(stderr,"sqlite_check_tables_with_id_obs_callback: foundID_OBS = %d but table_name is empty", foundID_OBS);
+      fprintf(stderr,"sqlite_check_tables_with_primary_key_callback: found_primary_key = %d but table_name is empty", found_primary_key);
       return 1;
     }
   }
 
   return 0;
-} /* End of function 'sqlite_check_tables_with_id_obs_callback' */
+} /* End of function 'sqlite_check_tables_with_primary_key_callback' */
 
 
   /***************************************************************************
-   * fonction: append_id_obs_table_list_requests
+   * fonction: append_primary_key_table_list_requests
    *
    * Cette fonction sert a ajouter des requetes SQL pour inclure les
-   * tables supplementaires autres que 'header' et 'data' mais qui ont
-   * 'id_obs' comme colonne.
+   * tables supplementaires autres que 'header' mais qui ont
+   * 'primary_key' comme colonne.
    *
    ***************************************************************************/
-void append_id_obs_table_list_requests(char* requete_sql, char* table_list) {
+void append_primary_key_table_list_requests(char* requete_sql, char* table_list, char* primary_key, char* header_table) {
   const char separator_char[2] = " ";
   char sqlreqtmp[MAXSTR], table_list_tmp[MAXSTR];
   char *token;
@@ -2652,12 +2693,13 @@ void append_id_obs_table_list_requests(char* requete_sql, char* table_list) {
   token = strtok(table_list_tmp, separator_char);
   /* walk through other tokens */
   while( token != (char*) NULL ) {
-    sprintf(sqlreqtmp,"insert into %s select * from dbin.%s where dbin.%s.id_obs in (select id_obs from header);\n",token,token,token);
+    sprintf(sqlreqtmp,"insert into %s select * from dbin.%s where dbin.%s.%s in (select %s from %s);\n",
+            token,token,token,primary_key,primary_key,header_table);
     strcat(requete_sql,sqlreqtmp);
     token = strtok((char*) NULL, separator_char);
   }
   // On a termine d'ajouter les requetes pour les autres tables
-}
+} /* End of function 'append_primary_key_table_list_requests' */
 
 
 /***************************************************************************
@@ -5071,6 +5113,27 @@ int parseOptions(int argc, char** argv, optionsptr optptr) {
 	}
 	optptr->roundrobin = 1;
       }
+      else if (!strcmp(argv[i],RDB_HEADER_OPTION)) { /* Cette option enregistrera le nom de la table qui joue le role du 'header' */
+	if (i+1>=argc || argv[i+1][0]=='-') {
+	  fprintf(stderr,"Fonction parseOptions: L'option %s demande un argument\n", argv[i]);
+	  exit_program(NOT_OK,PROGRAM_NAME,PROBLEM,VERSION);
+	}
+	strcpy(optptr->rdb_header_table,argv[++i]);
+      }
+      else if (!strcmp(argv[i],RDB_DATA_OPTION)) { /* Cette option enregistrera le nom de la table qui joue le role du 'data' */
+	if (i+1>=argc || argv[i+1][0]=='-') {
+	  fprintf(stderr,"Fonction parseOptions: L'option %s demande un argument\n", argv[i]);
+	  exit_program(NOT_OK,PROGRAM_NAME,PROBLEM,VERSION);
+	}
+	strcpy(optptr->rdb_data_table,argv[++i]);
+      }
+      else if (!strcmp(argv[i],RDB_PRIMARYKEY_OPTION)) { /* Cette option enregistrera le nom de la cle primaire qui lie 'header' et 'data' */
+	if (i+1>=argc || argv[i+1][0]=='-') {
+	  fprintf(stderr,"Fonction parseOptions: L'option %s demande un argument\n", argv[i]);
+	  exit_program(NOT_OK,PROGRAM_NAME,PROBLEM,VERSION);
+	}
+	strcpy(optptr->rdb_primarykey,argv[++i]);
+      }
       else if (!strcmp(argv[i],GZ_OPTION)) { /* Fichier standard dans lequel on lira le GZ au niveau voulu */
 	if (i+1>=argc || argv[i+1][0]=='-') {
 	  fprintf(stderr, "Fonction parseOptions: L'option %s demande un argument\n", argv[i]);
@@ -5350,8 +5413,13 @@ void aide(void) {
   printf("  %s   [active la verification a chaque enregistrement si on est en presence d'un UA multi-niveaux (ua4d)]\n\n", CHECK_UA4D_OPTION);
 
   printf("Attention, les options '%s' et '%s' puis '%s' et '%s' viennent en couple"
-	 " et sont mutuellement exclusives\n", RDBIN_OPTION, RDBOUT_OPTION, BURPIN_OPTION, BURPOUT_OPTION);
+	 " et sont mutuellement exclusives\n\n", RDBIN_OPTION, RDBOUT_OPTION, BURPIN_OPTION, BURPOUT_OPTION);
   
+  printf("On specifie les tables et la cle primaire qui lie les tables ensemble a l'aide des options suivantes\n");
+  printf("  %s     [table qui joue le role du 'header' (par defaut '%s')]\n", RDB_HEADER_OPTION, RDB_HEADER_DEFAUT);
+  printf("  %s       [table qui joue le role du 'data' (par defaut '%s')]\n", RDB_DATA_OPTION, RDB_DATA_DEFAUT);
+  printf("  %s [cle primaire qui lie les tables 'header' et 'data' (par defaut '%s')]\n\n", RDB_PRIMARYKEY_OPTION, RDB_PRIMARYKEY_DEFAUT);
+
   printf("Les options suivantes servent a identifier le champ qui definit la grille (fonction fstinf):\n\n");
 
   printf("  %s    [nomvar du champ qu'on veut lire] (par defaut, PN)\n", NOMVAR_OPTION);