+++ CHANGELOG.md
@@ -60,6 +60,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+ * Generalize the input table names and primary key for program `midas.splitobs.Abs` (#570 and !536)
  * Add MPI support over lat-bands for implicit diffusion solver (#436 and !533)
    * Minor impact on results (only when using implicit diffusion solver)
  * Support to read updated ObsDB SQLite observation tables and columns by use new ascii dictionary ObsDB - ObsSpaceData file (#528 and !524)